requirements.txt
# langchain
# langchain-community
# langchain-text-splitters
# langchain-ollama
# sentence-transformers

# pypdf
# streamlit
# faiss-cpu

streamlit>=1.42.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-huggingface>=0.1.0
langchain-groq>=0.2.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
torch>=2.4.0